implemented a session handler
