+++ package.json
@@ -15,6 +15,7 @@
   "license": "MIT",
   "dependencies": {
     "bunyan": "^1.8.14",
+    "cookie-session": "^1.4.0",
     "date-fns": "^2.16.1",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
